Update student/professor home to use Bootstrap flex grid system.
These changes enhance the spacing of the class cards regardless of
the number of available courses. It also enhances the look of the
cards themselves.

Also added styling to differentiate "form fields" and their values,
and space out information so that it makes more sense. Adjust the
layout of the cards so that certain elements like the class code,
title, and button row are aligned vertically along the different
class cards.